Create VariationQuickUpdateFill and move list price related menu items and groups to block
