+++ packages/js/product-editor/src/components/variations-table/pricing-menu-item/pricing-menu-item.tsx
@@ -13,47 +13,10 @@ import { chevronRight } from '@wordpress/icons';
 import { TRACKS_SOURCE } from '../../../constants';
 import { handlePrompt } from '../../../utils/handle-prompt';
 import { VariationActionsMenuItemProps } from '../types';
-import { SetListPriceMenuItem } from '../set-list-price-menu-item';
-import { VariationQuickUpdateMenuItem } from '../variation-actions-menus';
-
-function isPercentage( value: string ) {
-	return value.endsWith( '%' );
-}
-
-function parsePercentage( value: string ) {
-	const stringNumber = value.substring( 0, value.length - 1 );
-	if ( Number.isNaN( Number( stringNumber ) ) ) {
-		return undefined;
-	}
-	return Number( stringNumber );
-}
-
-function addFixedOrPercentage(
-	value: string,
-	fixedOrPercentage: string,
-	increaseOrDecrease: 1 | -1 = 1
-) {
-	if ( isPercentage( fixedOrPercentage ) ) {
-		if ( Number.isNaN( Number( value ) ) ) {
-			return 0;
-		}
-		const percentage = parsePercentage( fixedOrPercentage );
-		if ( percentage === undefined ) {
-			return Number( value );
-		}
-		return (
-			Number( value ) +
-			Number( value ) * ( percentage / 100 ) * increaseOrDecrease
-		);
-	}
-	if ( Number.isNaN( Number( value ) ) ) {
-		if ( Number.isNaN( Number( fixedOrPercentage ) ) ) {
-			return undefined;
-		}
-		return Number( fixedOrPercentage );
-	}
-	return Number( value ) + Number( fixedOrPercentage ) * increaseOrDecrease;
-}
+import {
+	VariationQuickUpdateFill,
+	VariationQuickUpdateMenuItem,
+} from '../variation-actions-menus';
 
 export function PricingMenuItem( {
 	selection,
@@ -87,100 +50,10 @@ export function PricingMenuItem( {
 			) }
 			renderContent={ () => (
 				<div className="components-dropdown-menu__menu">
-					<MenuGroup label={ __( 'List price', 'woocommerce' ) }>
-						<SetListPriceMenuItem
-							selection={ selection }
-							onChange={ onChange }
-							onClose={ onClose }
-						/>
-						<MenuItem
-							onClick={ () => {
-								recordEvent(
-									'product_variations_menu_pricing_select',
-									{
-										source: TRACKS_SOURCE,
-										action: 'list_price_increase',
-										variation_id: ids,
-									}
-								);
-								handlePrompt( {
-									message: __(
-										'Enter a value (fixed or %)',
-										'woocommerce'
-									),
-									onOk( value ) {
-										recordEvent(
-											'product_variations_menu_pricing_update',
-											{
-												source: TRACKS_SOURCE,
-												action: 'list_price_increase',
-												variation_id: ids,
-											}
-										);
-										onChange(
-											selection.map(
-												( { id, regular_price } ) => ( {
-													id,
-													regular_price:
-														addFixedOrPercentage(
-															regular_price,
-															value
-														)?.toFixed( 2 ),
-												} )
-											)
-										);
-									},
-								} );
-								onClose();
-							} }
-						>
-							{ __( 'Increase list price', 'woocommerce' ) }
-						</MenuItem>
-						<MenuItem
-							onClick={ () => {
-								recordEvent(
-									'product_variations_menu_pricing_select',
-									{
-										source: TRACKS_SOURCE,
-										action: 'list_price_decrease',
-										variation_id: ids,
-									}
-								);
-								handlePrompt( {
-									message: __(
-										'Enter a value (fixed or %)',
-										'woocommerce'
-									),
-									onOk( value ) {
-										recordEvent(
-											'product_variations_menu_pricing_update',
-											{
-												source: TRACKS_SOURCE,
-												action: 'list_price_increase',
-												variation_id: ids,
-											}
-										);
-										onChange(
-											selection.map(
-												( { id, regular_price } ) => ( {
-													id,
-													regular_price:
-														addFixedOrPercentage(
-															regular_price,
-															value,
-															-1
-														)?.toFixed( 2 ),
-												} )
-											)
-										);
-									},
-								} );
-								onClose();
-							} }
-						>
-							{ __( 'Decrease list price', 'woocommerce' ) }
-						</MenuItem>
-					</MenuGroup>
+					<VariationQuickUpdateFill.Slot
+						name="list-price"
+						fillProps={ { selection, onChange, onClose } }
+					/>
 					<MenuGroup label={ __( 'Sale price', 'woocommerce' ) }>
 						<MenuItem
 							onClick={ () => {
