+++ packages/js/product-editor/src/components/variations-table/variation-actions-menus/index.ts
@@ -1,4 +1,5 @@
 export * from './multiple-update-menu';
 export * from './single-update-menu';
 export * from './variation-quick-update-menu-item';
+export * from './variation-quick-update-fill';
 export * from './types';
